Fonction heuristique V1 finie (choix des batches par ordre de date due min croissante

+++ Client.cpp
@@ -1,13 +1,13 @@
 #include "Client.h"
 
-Client::Client(int h, int cout_u,int dist) {
+Client::Client(int h, float cout_u,int dist) {
 	fini = false;
 	this->h = h;
 	this->cout_u = cout_u;
 	this->dist = dist;
 }
 
-int Client::getCoutUnitaireStockage() {
+float Client::getCoutUnitaireStockage() {
 	return cout_u;
 }
